// src/app/model/profile.ts
export class Profile{
    id: number;
    name: string;
    profesion: string;
    picture: string;
    experience: number;
    email: string;
    website: string;
    phone: number;
    description: string;
    skill_1: string;
    valueSkill_1: number;
    skill_2: string;
    valueSkill_2: number;
    skill_3: string;
    valueSkill_3: number;
    skill_4: string;
    valueSkill_4: number;
    education: string;

    
  }